pyproject.toml
[project]
name = "ellellehm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.8.2",
    "nanoid>=2.0.0",
    "openai>=1.68.2",
    "pydantic>=2.10.6",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
]
